int->size_t in for
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist/trunk/OpenRTM-aist@2226 8bdb766d-82ee-4cb2-b2de-c55db819e4d5

+++ src/lib/coil/common/crc.cpp
@@ -85,7 +85,7 @@ namespace coil
         0x6E17, 0x7E36, 0x4E55, 0x5E74, 0x2E93, 0x3EB2, 0x0ED1, 0x1EF0,
       };
     unsigned short crc16(0xffff);
-    for (int i(0); i < len; ++i)
+    for (size_t i(0); i < len; ++i)
       {
         crc16 = crc16tab[(crc16 >> 8) ^ str[i]] ^ (crc16 << 8);
       }
